Add interactive features and optimize data pipeline
Frontend:
- Change spot callout to appear on hover instead of click
- Add chart date range selector component
- Improve map, menu sliders, and modal chart components
- Add utility functions

Backend:
- Consolidate 1D histograms into single file for faster loading
- Update histogram repository, service, and schema
- Improve spot and windrose services

Data Pipeline:
- Add histogram migration script
- Update CDS service and wind processor
- Optimize configuration

+++ backend/schemas/histogram.py
@@ -8,7 +8,7 @@ class HistogramResponse(BaseModel):
 
     spot_id: str
     bins: List[float]
-    daily_data: Dict[str, List[int]]  # {"01-01": [counts...], ...}
+    daily_data: Dict[str, List[float]]  # {"01-01": [counts...], ...}
 
 
 class WindRoseResponse(BaseModel):
